upd: separated donation and contact page
