+++ frontend/src/app/Donate/page.tsx
@@ -1,0 +1,9 @@
+import Donation from "../../components/Donation";
+
+export default function DonationPage() {
+  return (
+    <main style={{ display: "flex", height: "100vh", overflow: "hidden" }}>
+      <Donation />
+    </main>
+  );
+}
